Make Ctrl+Alt+w pipe widget information to console
The key combination used for this is likely to change. See
ConsoleWindow.designateCurrentWidget() for precisely what this does and
means.

Essentially, mousing over something (with the console open) and hitting
that combination of keys will give information on the GUI element the user
had the mouse over at the time (if any), via the console.

It also tells you what the stylesheet looks like, if it has one; if it
doesn't have one, it looks for one on the object's parents, posting that
instead if applicable.

The last selected widget is stored in CONSOLE.selected_widget - it can
be modified directly from there.

More features and detailed information may come in the future.